move origin to server config
