+++ packages/api/config/default.json
@@ -3,7 +3,8 @@
 		"playground": true,
 		"introspection": true,
 		"port": 3030,
-		"host": "localhost"
+		"host": "localhost",
+		"origin": "http://localhost"
 	},
 	"logging": {
 		"pretty": true
