+++ packages/api/config/demo.json
@@ -1,8 +1,10 @@
 {
 	"server": {
+		"playground": null,
+		"introspection": false,
 		"port": 8080,
 		"host": "0.0.0.0",
-		"origin": "https://cbo-ops-suite-integ.azureedge.net"
+		"origin": "https://cbo-ops-suite-demo.azureedge.net"
 	},
 	"logging": {
 		"pretty": false
